[step1] chore : 자바 8 스트림, 람다, optional 실습에 따른 코드 작성
