+++ src/main/java/step1/fp/Conditional.java
@@ -1,0 +1,5 @@
+package step1.fp;
+
+public interface Conditional {
+	boolean check(Integer number);
+}
