+++ src/main/java/step1/fp/StreamStudy.java
@@ -1,0 +1,53 @@
+package step1.fp;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class StreamStudy {
+
+    public static long countWords() throws IOException {
+        String contents = Files.readString(Paths
+                .get("src/main/resources/fp/war-and-peace.txt"));
+        List<String> words = Arrays.asList(contents.split("\\P{L}+"));
+
+        long count = words.stream().filter(w -> w.length() > 12).count();
+		return count;
+    }
+
+    public static void printLongestWordTop100() throws IOException {
+        String contents = Files.readString(Paths
+                .get("src/main/resources/fp/war-and-peace.txt"));
+        List<String> words = Arrays.asList(contents.split("\\P{L}+"));
+
+		List<String> collect = words.stream()
+				.filter(x -> x.length() > 12)
+				.sorted(Comparator.comparingInt(String::length))
+				.distinct()
+				.limit(100)
+				.map(String::toLowerCase)
+				.collect(Collectors.toList());
+
+		System.out.println(collect);
+    }
+
+    public static List<Integer> doubleNumbers(List<Integer> numbers) {
+        return numbers.stream().map(x -> 2 * x).collect(Collectors.toList());
+    }
+
+    public static long sumAll(List<Integer> numbers) {
+        return numbers.stream().reduce(0, (x, y) -> x + y);
+    }
+
+    public static long sumOverThreeAndDouble(List<Integer> numbers) {
+        return numbers.stream()
+				.filter(x -> x > 3)
+				.map(x -> x * 2)
+				.reduce(0, Integer::sum);
+    }
+}
