+++ src/main/java/step1/optional/Computer.java
@@ -1,4 +1,4 @@
-package nextstep.optional;
+package step1.optional;
 
 public class Computer {
     private Soundcard soundcard;
