+++ src/main/java/step1/optional/ComputerStore.java
@@ -1,7 +1,7 @@
-package nextstep.optional;
+package step1.optional;
 
-import nextstep.optional.Computer.Soundcard;
-import nextstep.optional.Computer.USB;
+import step1.optional.Computer.Soundcard;
+import step1.optional.Computer.USB;
 
 public class ComputerStore {
     public static final String UNKNOWN_VERSION = "UNKNOWN";
