+++ src/main/java/step1/optional/Expression.java
@@ -1,4 +1,6 @@
-package nextstep.optional;
+package step1.optional;
+
+import java.util.Arrays;
 
 enum Expression {
     PLUS("+"), MINUS("-"), TIMES("*"), DIVIDE("/");
@@ -14,12 +16,9 @@ enum Expression {
     }
 
     static Expression of(String expression) {
-        for (Expression v : values()) {
-            if (matchExpression(v, expression)) {
-                return v;
-            }
-        }
-
-        throw new IllegalArgumentException(String.format("%s는 사칙연산에 해당하지 않는 표현식입니다.", expression));
+		return Arrays.stream(values())
+				.filter(v -> matchExpression(v, expression))
+				.findFirst()
+				.orElseThrow(() -> new IllegalArgumentException(String.format("%s는 사칙연산에 해당하지 않는 표현식입니다.", expression)));
     }
 }
