+++ src/test/java/step1/fp/StreamStudyTest.java
@@ -1,4 +1,4 @@
-package nextstep.fp;
+package step1.fp;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
