+++ src/test/java/step1/optional/UserTest.java
@@ -1,9 +1,9 @@
-package nextstep.optional;
+package step1.optional;
 
 import org.junit.jupiter.api.Test;
 
-import static nextstep.optional.User.ageIsInRange1;
-import static nextstep.optional.User.ageIsInRange2;
+import static step1.optional.User.ageIsInRange1;
+import static step1.optional.User.ageIsInRange2;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class UserTest {
